refactor: centralize canonical network transaction planning
