+++ docs/worklog/2026-07-14-network-transaction-observation-contract.md
@@ -1,0 +1,120 @@
+# Network Transaction and Observation Contract
+
+## Status
+
+Feature branch: `codex/network-transaction-observation-contract`
+
+Base: `origin/dev` at `a0177422` after PR #356 merged. This work is divided into three
+reviewable milestones. Each milestone receives focused verification plus an
+`iteration-test-expanded` generation/evaluation and blind panel before the next milestone begins.
+
+## Family Contract
+
+### Owning abstractions
+
+- The network-connection action bundle owns the canonical connection occurrence: tuple, outcome,
+  phase anchors, connection state/history, duration, process ownership, and directional traffic
+  accounting.
+- Canonical `NetworkContext` and `StateManager` connection state own durable shared traffic truth.
+- Network visibility and observation planning own the sensor-visible tuple, sensor identity,
+  capture timing, packet loss, and source-local identifiers.
+- The dispatcher owns final source-visible output-window admission using typed action lifecycle
+  metadata.
+- The explicit-proxy action bundle owns the client/proxy/DNS/origin phase graph and proxy byte
+  semantics.
+- Emitters own only source-native serialization of already-planned truth.
+
+### Invariants
+
+1. One finalized canonical ledger supplies payload bytes, packet counts, IP bytes, duration,
+   state, and history to every consumer; action-generated fields are not independently mutated
+   after finalization.
+2. Lossless observations of one transaction retain identical accounting. NAT changes the visible
+   tuple only. Counter divergence requires an explicit capture-loss decision with positive
+   `missed_bytes`.
+3. Sensor clocks and route delays are persistent scoped properties plus bounded capture jitter,
+   not independently resampled broad offsets in emitters.
+4. A fresh action group whose final source-visible start is outside the half-open output window is
+   not rendered. Only typed closure evidence for a group that began before the end may trail past
+   the window.
+5. Proxy cache/deny/failure outcomes emit only phases that actually occurred. DNS and origin
+   transport timing come from a conditional phase graph without a universal delay floor.
+6. Proxy request/response body sizes and client/server transferred byte totals are distinct owned
+   values.
+
+### Entry paths
+
+- Baseline and storyline calls to `ActivityGenerator.generate_connection()`.
+- DNS, DHCP, SSH, RDP, browser, scanner, file-transfer, Kerberos, Windows remote-admin, and proxy
+  action bundles that compose the canonical connection entrypoint.
+- Causal prerequisite DNS and application-side-effect expansion.
+- Persistent HTTP connections and explicit-proxy tunnel reuse.
+- Direct canonical `SecurityEvent` construction used by focused tests and explicit low-level
+  compatibility paths.
+
+### Consumers
+
+- Zeek conn/protocol/file emitters and their per-sensor multiplexer.
+- Cisco ASA, Snort/Suricata, proxy, web, Sysmon, Windows Security, and eCAR FLOW renderers.
+- `StateManager` open-connection/session/process lifetime checks.
+- Source timing, observation policy, network visibility, NAT routing, eval parsers, hard probes,
+  and blind network/host/detection/hunting review.
+
+### Layer rationale
+
+The repeated findings cross source and emitter boundaries, so no renderer can own the repair.
+Connection truth must be complete at the action/context/state boundary; sensor-specific divergence
+must be explicit at the routing/observation boundary; admission must happen only after final source
+timing is known; and multi-connection proxy causality must stay in the higher proxy bundle.
+
+### Sibling risks and exclusions
+
+- Preserve source-native schema differences while removing invented semantic differences.
+- Preserve valid endpoint attribution/lifetime clamps for SSH, RDP, Kerberos, and proxy paths.
+- Keep direct test/raw events compatible when no canonical ledger or lifecycle metadata is
+  supplied; they receive a validated compatibility projection rather than failing outright.
+- Durable general-purpose thread, identity, role/activity, and session-bootstrap improvements are
+  separate assessment families unless this contract directly owns the observed network evidence.
+
+## Milestone Record
+
+### Milestone 1 — Canonical transaction and ledger
+
+Implementation and expanded-scenario acceptance complete:
+
+- Moved the full connection orchestration body from `ActivityGenerator` into the action-owned
+  `NetworkTransactionPlanner` without changing deterministic network behavior. The planner uses a
+  mutable action-owned occurrence draft, freezes tuple, hostname, process ownership, phase anchors,
+  outcome, state/history, interval, and accounting, and only then constructs `SecurityEvent`.
+- Added immutable directional/bidirectional traffic ledgers and a finalized transaction plan.
+- Finalized and validated action-generated `NetworkContext` truth immediately before dispatch;
+  direct/raw compatibility contexts continue to project a read-only ledger on demand.
+- Persisted the same ledger, interval, state/history, and duration in `OpenConnection`.
+- Persistent HTTP application transactions reuse the parent's connection identity and accumulate
+  its durable ledger. They reserve their prior logical request identity without opening shadow
+  state, preserving the established deterministic RNG/ID stream.
+- Added central invariant tests plus a non-sample IDS fan-out assertion.
+
+Verification:
+
+- Focused network/proxy/endpoint regression set: 332 passed, 1 skipped.
+- Default suite: 4,918 passed and 19 skipped; the two initial failures were separately verified:
+  the deterministic injection twin passed after preserving identity reservations, and the Splunk
+  harness passed outside the restricted socket sandbox.
+- Repository-wide Ruff lint and format checks passed.
+- `iteration-test-expanded` generated 93,351 records and passed automated evaluation at
+  95.83901668531391, exactly preserving the pre-milestone record count and score.
+- The independent panel scored 82, 69, 87, and 76 synthetic confidence (78.5 average). Its two
+  network-wide findings were the expected next contracts: emitter-owned per-flow sensor
+  timing/accounting differences and the proxy's narrow multi-second DNS/origin delay. Strong
+  transaction, protocol-parent, endpoint, TLS/certificate, and firewall joins were preserved.
+- No contract-level regression required a repeated panel. The ignored assessment bundle is under
+  `scenarios/iteration-test-expanded/blind-test/network-contract-milestone-1/`.
+
+### Milestone 2 — Sensor observation and final-window admission
+
+Pending.
+
+### Milestone 3 — Proxy phase and byte contract
+
+Pending.
