+++ src/evidenceforge/events/contexts.py
@@ -33,6 +33,13 @@ from dataclasses import dataclass, field
 from datetime import datetime
 from typing import Any
 
+from evidenceforge.events.network import (
+    DirectionalTrafficLedger,
+    NetworkTrafficLedger,
+    NetworkTransactionOutcome,
+    NetworkTransactionPlan,
+)
+
 
 @dataclass(slots=True)
 class HostContext:
@@ -181,6 +188,118 @@ class NetworkContext:
     responding_pid: int = -1  # Destination-side PID that accepted/owned the connection
     link_local: bool = False  # True for same-broadcast-domain traffic such as DHCP
     application_layer_only: bool = False  # Additional protocol transaction on an existing flow
+    transaction: NetworkTransactionPlan | None = None
+
+    @property
+    def traffic_ledger(self) -> NetworkTrafficLedger:
+        """Return finalized traffic truth or a compatibility projection."""
+
+        if self.transaction is not None:
+            return self.transaction.traffic
+        return self._project_traffic_ledger()
+
+    def finalize_transaction(
+        self,
+        stable_id: str,
+        *,
+        hostname: str = "",
+        outcome: NetworkTransactionOutcome = "success",
+        phase_times: tuple[tuple[str, datetime], ...] = (),
+    ) -> NetworkTransactionPlan:
+        """Freeze the final canonical transaction after all planning adjustments."""
+
+        started_at = self.source_visible_start_time
+        if started_at is None:
+            raise ValueError("Cannot finalize a network transaction without a visible start")
+        closed_at = self.source_visible_close_time
+        transaction = NetworkTransactionPlan(
+            stable_id=stable_id,
+            hostname=hostname,
+            outcome=outcome,
+            phase_times=phase_times or (("transport_start", started_at),),
+            started_at=started_at,
+            closed_at=closed_at,
+            src_ip=self.src_ip,
+            src_port=self.src_port,
+            dst_ip=self.dst_ip,
+            dst_port=self.dst_port,
+            protocol=self.protocol,
+            service=self.service,
+            zeek_uid=self.zeek_uid,
+            conn_id=self.conn_id,
+            duration=self.duration,
+            conn_state=self.conn_state,
+            history=self.history,
+            traffic=self._project_traffic_ledger(),
+            initiating_pid=self.initiating_pid,
+            responding_pid=self.responding_pid,
+        )
+        self.transaction = transaction
+        return transaction
+
+    def validate_finalized_transaction(self) -> None:
+        """Fail if compatibility fields drift from finalized canonical truth."""
+
+        transaction = self.transaction
+        if transaction is None:
+            return
+        projection = (
+            self.src_ip,
+            self.src_port,
+            self.dst_ip,
+            self.dst_port,
+            self.protocol,
+            self.service,
+            self.zeek_uid,
+            self.conn_id,
+            self.duration,
+            self.conn_state,
+            self.history,
+            self.initiating_pid,
+            self.responding_pid,
+            self.source_visible_start_time,
+            self.source_visible_close_time,
+            self._project_traffic_ledger(),
+        )
+        canonical = (
+            transaction.src_ip,
+            transaction.src_port,
+            transaction.dst_ip,
+            transaction.dst_port,
+            transaction.protocol,
+            transaction.service,
+            transaction.zeek_uid,
+            transaction.conn_id,
+            transaction.duration,
+            transaction.conn_state,
+            transaction.history,
+            transaction.initiating_pid,
+            transaction.responding_pid,
+            transaction.started_at,
+            transaction.closed_at,
+            transaction.traffic,
+        )
+        if projection != canonical:
+            raise ValueError("NetworkContext changed after canonical transaction finalization")
+
+    def _project_traffic_ledger(self) -> NetworkTrafficLedger:
+        """Build immutable accounting from legacy flat context fields."""
+
+        orig_payload = max(0, self.orig_bytes or 0)
+        resp_payload = max(0, self.resp_bytes or 0)
+        orig_packets = max(0, self.orig_pkts)
+        resp_packets = max(0, self.resp_pkts)
+        orig_ip_bytes = max(orig_payload, self.orig_ip_bytes or 0)
+        resp_ip_bytes = max(resp_payload, self.resp_ip_bytes or 0)
+        if orig_ip_bytes > 0 and orig_packets == 0:
+            orig_packets = 1
+        if resp_ip_bytes > 0 and resp_packets == 0:
+            resp_packets = 1
+        return NetworkTrafficLedger(
+            orig=DirectionalTrafficLedger(orig_payload, orig_packets, orig_ip_bytes),
+            resp=DirectionalTrafficLedger(resp_payload, resp_packets, resp_ip_bytes),
+            missed_orig_bytes=max(0, self.missed_bytes),
+        )
 
 
 @dataclass(slots=True)
